create a global provider & centralize error handling
