+++ app/src/app/components/GlobalProvider.tsx
@@ -1,0 +1,40 @@
+"use client";
+
+import { IFeature } from "@groceries/shared";
+import { MantineProvider } from "@mantine/core";
+import { useState, useEffect } from "react";
+import { FeaturesContext, FeatureSet, SetFeaturesContext } from "../../context/featuresContext";
+import { getFeatures } from "../../services/api";
+import ErrorHandler from "./ErrorHandler";
+import { SetErrorContext } from "../../context/errorContext";
+
+export default function GlobalProvider({ children }: { children: React.ReactNode }) {
+  const [features, setFeatures] = useState<FeatureSet>({});
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    getFeatures()
+      .then(data => {
+        // turn array of features into a map
+        setFeatures(data.reduce((acc: FeatureSet, f: IFeature) => {
+          acc[f.name] = f.enabled;
+          return acc;
+        }, {}));
+      })
+      .catch(_ => setError("Failed to get features... try again later"));
+  }, []);
+
+  return (
+    <MantineProvider defaultColorScheme="auto" >
+      <ErrorHandler error={error}>
+        <SetErrorContext.Provider value={setError}>
+          <FeaturesContext.Provider value={features}>
+            <SetFeaturesContext.Provider value={setFeatures}>
+              <div id="root">{children}</div>
+            </SetFeaturesContext.Provider>
+          </FeaturesContext.Provider>
+        </SetErrorContext.Provider>
+      </ErrorHandler>
+    </MantineProvider>
+  );
+}
