+++ app/src/context/errorContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const SetErrorContext = createContext<(error: string) => void >(() => {throw new Error("SetErrorContext not provided")} );
